Refactor delete duplicates script and consolidate it into one file
rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
